+++ src/AngularSPAWebAPI/app/main-aot.ts
@@ -1,6 +1,6 @@
 // AoT compilation entry point.
 
-// Polyfills (no need reflect-metadata).
+// Polyfills.
 import 'core-js/client/shim';
 import 'zone.js/dist/zone';
 
@@ -9,7 +9,7 @@ import 'web-animations-js';
 import 'hammerjs';
 
 // Styles.
-import './styles/styles.scss';
+import './styles.scss';
 
 import { platformBrowser } from '@angular/platform-browser';
 import { enableProdMode } from '@angular/core';