+++ src/AngularSPAWebAPI/app/resources/resources.component.ts
@@ -3,7 +3,8 @@
 import { AuthHttp } from 'angular2-jwt';
 
 @Component({
-    templateUrl: './resources.component.html'
+    templateUrl: './resources.component.html',
+    styleUrls: ['./resources.component.scss']
 })
 export class ResourcesComponent implements OnInit {